+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2:
 
+* Add RDate.addDays utility method.
 * Extract DateUtil from GwtDateTimeService.
 
 ## 0.1: